+++ app/src/main/java/woowacourse/shopping/di/sangoonDi/StartInjection.kt
@@ -1,6 +1,4 @@
-package woowacourse.shopping.di
-
-import android.util.Log
+package woowacourse.shopping.di.sangoonDi
 
 object StartInjection {
     operator fun invoke(block: StartInjection.() -> Unit) {
@@ -9,7 +7,6 @@ object StartInjection {
 
     inline fun <reified T : Any> single(instance: Any) {
         DependencyContainer.add(T::class, instance)
-        Log.d("123123", instance.toString())
     }
 
     // Disposable의 경우도 처리 필요
