+++ app/src/main/java/woowacourse/shopping/ui/main/MainActivity.kt
@@ -8,7 +8,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProvider
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityMainBinding
-import woowacourse.shopping.di.inject
+import woowacourse.shopping.di.sangoonDi.inject
 import woowacourse.shopping.ui.cart.CartActivity
 import woowacourse.shopping.ui.main.adapter.ProductAdapter
